# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "M-Eyes",
    "description": "Open-source DDI platform (DNS, DHCP, IPAM) with Fortinet ecosystem integration",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/dhcp/subnets": {
      "get": {
        "tags": [
          "dhcp"
        ],
        "summary": "List Subnets",
        "operationId": "list_subnets_api_v1_dhcp_subnets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DhcpSubnetOut"
                  },
                  "type": "array",
                  "title": "Response List Subnets Api V1 Dhcp Subnets Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "dhcp"
        ],
        "summary": "Create Subnet",
        "operationId": "create_subnet_api_v1_dhcp_subnets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DhcpSubnetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DhcpSubnetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/dhcp/subnets/{subnet_id}": {
      "get": {
        "tags": [
          "dhcp"
        ],
        "summary": "Get Subnet",
        "operationId": "get_subnet_api_v1_dhcp_subnets__subnet_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subnet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Subnet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DhcpSubnetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "dhcp"
        ],
        "summary": "Update Subnet",
        "operationId": "update_subnet_api_v1_dhcp_subnets__subnet_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subnet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Subnet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DhcpSubnetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DhcpSubnetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "dhcp"
        ],
        "summary": "Delete Subnet",
        "operationId": "delete_subnet_api_v1_dhcp_subnets__subnet_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subnet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Subnet Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dhcp/subnets/{subnet_id}/ranges": {
      "post": {
        "tags": [
          "dhcp"
        ],
        "summary": "Create Range",
        "operationId": "create_range_api_v1_dhcp_subnets__subnet_id__ranges_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subnet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Subnet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DhcpRangeIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DhcpRangeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dhcp/ranges/{range_id}": {
      "delete": {
        "tags": [
          "dhcp"
        ],
        "summary": "Delete Range",
        "operationId": "delete_range_api_v1_dhcp_ranges__range_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "range_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Range Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dhcp/subnets/{subnet_id}/reservations": {
      "post": {
        "tags": [
          "dhcp"
        ],
        "summary": "Create Reservation",
        "operationId": "create_reservation_api_v1_dhcp_subnets__subnet_id__reservations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subnet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Subnet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DhcpReservationIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DhcpReservationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dhcp/reservations/{reservation_id}": {
      "delete": {
        "tags": [
          "dhcp"
        ],
        "summary": "Delete Reservation",
        "operationId": "delete_reservation_api_v1_dhcp_reservations__reservation_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reservation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dhcp/options": {
      "get": {
        "tags": [
          "dhcp"
        ],
        "summary": "List Global Options",
        "operationId": "list_global_options_api_v1_dhcp_options_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DhcpOptionOut"
                  },
                  "type": "array",
                  "title": "Response List Global Options Api V1 Dhcp Options Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "dhcp"
        ],
        "summary": "Set Global Option",
        "operationId": "set_global_option_api_v1_dhcp_options_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DhcpOptionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DhcpOptionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/dhcp/subnets/{subnet_id}/options": {
      "post": {
        "tags": [
          "dhcp"
        ],
        "summary": "Set Subnet Option",
        "operationId": "set_subnet_option_api_v1_dhcp_subnets__subnet_id__options_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subnet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Subnet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DhcpOptionIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DhcpOptionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dhcp/options/{option_id}": {
      "delete": {
        "tags": [
          "dhcp"
        ],
        "summary": "Delete Option",
        "operationId": "delete_option_api_v1_dhcp_options__option_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "option_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Option Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DhcpOptionIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "name",
          "value"
        ],
        "title": "DhcpOptionIn"
      },
      "DhcpOptionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "subnet_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subnet Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "subnet_id",
          "name",
          "value"
        ],
        "title": "DhcpOptionOut"
      },
      "DhcpRangeIn": {
        "properties": {
          "start_ip": {
            "type": "string",
            "title": "Start Ip"
          },
          "end_ip": {
            "type": "string",
            "title": "End Ip"
          }
        },
        "type": "object",
        "required": [
          "start_ip",
          "end_ip"
        ],
        "title": "DhcpRangeIn"
      },
      "DhcpRangeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "subnet_id": {
            "type": "integer",
            "title": "Subnet Id"
          },
          "start_ip": {
            "type": "string",
            "title": "Start Ip"
          },
          "end_ip": {
            "type": "string",
            "title": "End Ip"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "subnet_id",
          "start_ip",
          "end_ip"
        ],
        "title": "DhcpRangeOut"
      },
      "DhcpReservationIn": {
        "properties": {
          "mac": {
            "type": "string",
            "title": "Mac"
          },
          "ip": {
            "type": "string",
            "title": "Ip"
          },
          "hostname": {
            "type": "string",
            "title": "Hostname",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "mac",
          "ip"
        ],
        "title": "DhcpReservationIn"
      },
      "DhcpReservationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "subnet_id": {
            "type": "integer",
            "title": "Subnet Id"
          },
          "mac": {
            "type": "string",
            "title": "Mac"
          },
          "ip": {
            "type": "string",
            "title": "Ip"
          },
          "hostname": {
            "type": "string",
            "title": "Hostname"
          },
          "ip_address_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "subnet_id",
          "mac",
          "ip",
          "hostname",
          "ip_address_id"
        ],
        "title": "DhcpReservationOut"
      },
      "DhcpSubnetIn": {
        "properties": {
          "network_id": {
            "type": "integer",
            "title": "Network Id"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "network_id"
        ],
        "title": "DhcpSubnetIn"
      },
      "DhcpSubnetOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "network_id": {
            "type": "integer",
            "title": "Network Id"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "cidr": {
            "type": "string",
            "title": "Cidr",
            "default": ""
          },
          "ranges": {
            "items": {
              "$ref": "#/components/schemas/DhcpRangeOut"
            },
            "type": "array",
            "title": "Ranges",
            "default": []
          },
          "reservations": {
            "items": {
              "$ref": "#/components/schemas/DhcpReservationOut"
            },
            "type": "array",
            "title": "Reservations",
            "default": []
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/DhcpOptionOut"
            },
            "type": "array",
            "title": "Options",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "network_id",
          "enabled"
        ],
        "title": "DhcpSubnetOut"
      },
      "DhcpSubnetUpdate": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "DhcpSubnetUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}